Refactored files to separate drawing and animation from main; Added bomb animation
